+++ sync-for-payables/CodatSyncPayables/Models/Operations/ListPaymentMethodsRequest.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncPayables.Models.Operations
+namespace Codat.Sync.Payables.Models.Operations
 {
-    using CodatSyncPayables.Utils;
+    using Codat.Sync.Payables.Utils;
     
     public class ListPaymentMethodsRequest
     {