+++ sync-for-payables/RELEASES.md
@@ -87,3 +87,13 @@ Based on:
 - [csharp v2.3.0] sync-for-payables
 ### Releases
 - [NuGet v2.3.0] https://www.nuget.org/packages/Codat.Sync.Payables/2.3.0 - sync-for-payables
+
+## 2023-11-13 01:28:36
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.0] sync-for-payables
+### Releases
+- [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Sync.Payables/3.0.0 - sync-for-payables